refactor: Move config folder to root
