+++ tsconfig.json
@@ -30,7 +30,7 @@
     "moduleResolution": "node",                          /* Specify how TypeScript looks up a file from a given module specifier. */
     "baseUrl": "./src",                                  /* Specify the base directory to resolve non-relative module names. */
     "paths": {
-      "@coreConfig/*": ["core/config/*"],
+      "@config/*": ["config/*"],
       "@coreModules/*": ["core/modules/*"],
       "@coreTypes/*": ["core/types/*"],
       "@coreUtils/*": ["core/utils/*"],
